move tree context to lib
